requirements.txt
Pillow==11.0.0
PyYAML==6.0.2
pytest==8.3.2
